% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{aggregate_classes}
\alias{aggregate_classes}
\title{Aggregate Classes in a Contingency Matrix}
\usage{
aggregate_classes(conf_matrix, classes_to_aggregate)
}
\arguments{
\item{conf_matrix}{A contingency matrix with class labels as row and column names.}

\item{classes_to_aggregate}{A vector of class labels to be aggregated into a single class.}
}
\value{
A contingency matrix with the specified classes aggregated.
}
\description{
This function aggregates specified classes in a contingency matrix into a single class.
}
\examples{
# Creating a sample contingency matrix
confusion_matrix <- matrix(
 c(285, 3, 1, 86, 41,
   0, 488, 74, 0, 1,
   0, 1386, 223, 1, 17,
   1, 51, 7, 86, 15,
   22, 357, 73, 64, 185),
 nrow = 5, byrow = TRUE,
 dimnames = list(
   Prediction = c("A", "B", "C", "D", "E"),
   Reference = c("A", "B", " C", "D", "E")
 )
)
aggregated_conf_matrix <- aggregate_classes(confusion_matrix, c("D", "E"))
}
