% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{ball_graphs}
\alias{ball_graphs}
\title{Create Bubble Plot from Confusion Matrix}
\usage{
ball_graphs(confusion_matrix, title = NULL, dict = NULL)
}
\arguments{
\item{confusion_matrix}{A confusion matrix with class labels as row and column names.}

\item{title}{A character string specifying the title of the plot.}

\item{dict}{An optional named vector for renaming the class labels.}
}
\value{
A ggplot object representing the bubble plot.
}
\description{
This function generates a bubble plot from a confusion matrix, with options for custom titles and label dictionaries.
}
\examples{
library(reshape2)
library(dplyr)
library(ggplot2)
conf_matrix <- matrix(c(50, 10, 5, 2, 30, 15, 3, 2, 40),
                      nrow = 3, byrow = TRUE,
                      dimnames = list(Prediction = c("Class1", "Class2", "Class3"),
                                      Reference = c("Class1", "Class2", "Class3")))
dict <- c("Class1" = "A", "Class2" = "B", "Class3" = "C")
ball_graphs(conf_matrix, title = "Confusion Matrix Bubble Plot", dict = dict)


}
