% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_fitting_test_KS.R
\name{best_ENL}
\alias{best_ENL}
\title{Find the Best ENL Based on Gamma Fitting Results}
\usage{
best_ENL(fitting_result_df)
}
\arguments{
\item{fitting_result_df}{A data frame containing the results of Gamma fitting, including columns for Class, Alpha, Beta, p_value, and H0.}
}
\value{
A numeric value representing the best ENL (alpha) or NA if no class accepted the null hypothesis.
}
\description{
This function finds the best Equivalent Number of Looks (ENL) based on the results of fitting a Gamma distribution. It returns the alpha parameter of the class with the highest p-value that accepted the null hypothesis.
}
\examples{
fitting_result_df <- data.frame(Class = c("Class1", "Class2"),
                                Alpha = c(2.5, 3.0),
                                Beta = c(1.5, 2.0),
                                p_value = c(0.05, 0.2),
                                H0 = c("reject", "accept"))
best_enl <- best_ENL(fitting_result_df)
print(best_enl)
}
