% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{bhattacharyya_distance}
\alias{bhattacharyya_distance}
\title{Calculate Bhattacharyya Distance Between Two Classes}
\usage{
bhattacharyya_distance(
  df_samples,
  class1,
  class2,
  p_function = "gaussian",
  L = NULL
)
}
\arguments{
\item{df_samples}{A data frame containing samples with a column named 'Class' indicating the class of each sample.}

\item{class1}{A character string specifying the label of the first class.}

\item{class2}{A character string specifying the label of the second class.}

\item{p_function}{A character string specifying the probability distribution function to be used ('gaussian' or 'gamma'). Default is 'gaussian'.}

\item{L}{An optional numeric value used in the calculation of the Bhattacharyya distance for the 'gamma' distribution.}
}
\value{
A numeric value representing the Bhattacharyya distance between the two classes.
}
\description{
This function calculates the Bhattacharyya distance between two classes in a data frame of samples using either a Gaussian or Gamma distribution.
}
\examples{
df_samples <- data.frame(Class = rep(c("Class1", "Class2"), each = 100),
                         Feature1 = rnorm(200), Feature2 = rnorm(200))
distance_gaussian <- bhattacharyya_distance(df_samples, "Class1", "Class2", p_function = 'gaussian')

df_samples_gamma <- data.frame(Class = rep(c("Class1", "Class2"), each = 100),
                               Feature1 = rgamma(200, shape = 2, rate = 1))
distance_gamma <- bhattacharyya_distance(df_samples_gamma, "Class1", "Class2", p_function = 'gamma', L = 2)
}
