% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_CMAP_functions.R
\name{calcularTransicoesImpossiveis}
\alias{calcularTransicoesImpossiveis}
\title{Calculate Impossible Transitions in a Time Series of Raster Classifications}
\usage{
calcularTransicoesImpossiveis(diretorioBase, anos, caminhosTransicoes)
}
\arguments{
\item{diretorioBase}{The base directory where the raster files are located.}

\item{anos}{A vector of years corresponding to the raster files.}

\item{caminhosTransicoes}{A vector of file paths to the transition matrices.}
}
\value{
The percentage of impossible transitions.
}
\description{
This function calculates the percentage of impossible transitions in a time series of raster classifications based on transition matrices.
}
\examples{
# Assuming example classification rasters are saved in "Classification_2000.TIF", "Classification_2001.TIF"
diretorioBase <- "path/to/directory/"
anos <- c(2000, 2001)
caminhosTransicoes <- c("path/to/transition1.txt", "path/to/transition2.txt")
if (all(file.exists(paste0(diretorioBase, "Classification_", anos, ".TIF"))) && all(file.exists(caminhosTransicoes))) {
  porcentagemImpossiveis <- calcularTransicoesImpossiveis(diretorioBase, anos, caminhosTransicoes)
} else {
  message("Example files do not exist. Please provide the correct paths to the classification rasters and transition matrices.")
}
}
