% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{class_generalization}
\alias{class_generalization}
\title{Generalize Classes in a Data Frame of Samples}
\usage{
class_generalization(df_samples, hierarchical_levels, level = 1)
}
\arguments{
\item{df_samples}{A data frame containing samples with a column named 'Class' indicating the class of each sample.}

\item{hierarchical_levels}{A list of hierarchical levels, where each level is a list of classes to be generalized.}

\item{level}{An integer indicating the hierarchical level to use for generalization. Default is 1.}
}
\value{
A data frame with generalized class labels.
}
\description{
This function generalizes classes in a data frame of samples based on hierarchical levels.
}
\examples{
df_samples <- data.frame(Class = c("Class1", "Class2", "Class3"), Value = rnorm(3))
hierarchical_levels <- list(list(c("Class1", "Class2"), c("Class3")))
generalized_df <- class_generalization(df_samples, hierarchical_levels, level = 1)

}
