% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_CMAP_functions.R
\name{CMAP_classifier}
\alias{CMAP_classifier}
\title{Compound Maximum A Posteriori (CMAP) Classifier}
\usage{
CMAP_classifier(
  class_likelihood_paths,
  traj_weights,
  output_path,
  masking = NULL
)
}
\arguments{
\item{class_likelihood_paths}{A character vector containing the file paths of the class likelihood raster images for each date.}

\item{traj_weights}{A data frame containing the trajectory weights and the sequences of class transitions.}

\item{output_path}{A character string specifying the directory where the output classified images will be saved.}

\item{masking}{An optional character vector containing file paths of masking images for each date. Default is NULL.}
}
\value{
A matrix representing the classification results for each date.
}
\description{
This function implements the CMAP classifier, which uses class likelihoods and trajectory weights to classify remote sensing images.
}
