% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_confusion_matrix.R
\name{cross_validation}
\alias{cross_validation}
\title{Cross Validation for Classification}
\usage{
cross_validation(
  df_samples,
  classes,
  sample_proportion = 0.7,
  p_function = "gaussian",
  n_iteracoes = 10,
  ENL = NULL,
  seed = NULL,
  aggregation = NULL,
  external_valid_samples = NULL,
  n_samples = NULL,
  fill_names = NULL,
  type = "variable"
)
}
\arguments{
\item{df_samples}{A data frame containing samples with a column named 'Class' indicating the class of each sample.}

\item{classes}{A vector of class labels to include in the validation process.}

\item{sample_proportion}{A numeric value representing the proportion of samples to be used for training. Default is 0.7.}

\item{p_function}{A character string specifying the probability distribution function to be used ('gaussian', 'gamma', or 'intensity_joint_distribution'). Default is 'gaussian'.}

\item{n_iteracoes}{An integer value specifying the number of iterations for cross validation. Default is 10.}

\item{ENL}{An optional numeric value used in the estimation for the 'gamma' and 'intensity_joint_distribution' distributions.}

\item{seed}{An optional numeric value or vector of seeds for reproducibility. If provided, the length of the vector determines the number of iterations.}

\item{aggregation}{An optional list of class aggregations to be applied to the confusion matrix.}

\item{external_valid_samples}{An optional data frame containing external validation samples.}

\item{n_samples}{An optional numeric value specifying the number of samples per class for validation.}

\item{fill_names}{An optional character vector for renaming the classes in the confusion matrix.}

\item{type}{A character string indicating the type of cross validation ('variable' or 'fixed'). Default is 'variable'.}
}
\value{
A list containing the overall accuracy, F1 macro scores, producers' accuracy, users' accuracy, F1 scores, and the seeds for the best, median, and third quartile accuracies.
}
\description{
This function performs cross validation on a dataset of samples to evaluate the performance of different classifiers.
}
\examples{
# Example usage of cross_validation function
df_samples <- data.frame(Class = rep(c("A", "B"), each = 25), matrix(rnorm(200), nrow = 50))
colnames(df_samples) <- c("Class", paste0("Feature", 1:4))
classes <- c("A", "B")
result <- cross_validation(df_samples, classes, sample_proportion = 0.7,
                           p_function = 'gaussian', n_iteracoes = 10, seed = NULL, type = 'variable')
print(result)
}
