% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{data_analysis}
\alias{data_analysis}
\title{Perform Data Analysis and Plotting}
\usage{
data_analysis(
  df_samples,
  classes,
  p_function = "gaussian",
  L = NULL,
  title = NULL
)
}
\arguments{
\item{df_samples}{A data frame containing the samples and their classes.}

\item{classes}{A vector of class labels.}

\item{p_function}{The probability function to use ('gaussian' or 'gamma').}

\item{L}{Optional parameter for the 'gamma' function.}

\item{title}{The title for the plots.}
}
\value{
The separability metric.
}
\description{
This function performs data analysis and generates plots based on the input data and specified parameters.
}
\examples{
df_samples <- data.frame(Class = rep(c("Class1", "Class2"), each = 50), Value = rnorm(100))
classes <- c("Class1", "Class2")
separability <- data_analysis(df_samples, classes, p_function = 'gaussian', title = "Amplitude")
}
