% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{df2img}
\alias{df2img}
\title{Convert a Data Frame to an Image Array}
\usage{
df2img(df_img, image)
}
\arguments{
\item{df_img}{A data frame where each column represents a band of the image.}

\item{image}{A template image from which the spatial dimensions and CRS are taken.}
}
\value{
A 3D image array reconstructed from the data frame.
}
\description{
This function converts a data frame back into a 3D image array, maintaining the original spatial dimensions and bands.
}
\examples{
library(raster)
df_img <- data.frame(Band1 = runif(100), Band2 = runif(100), Band3 = runif(100))
template_raster <- stack(replicate(3, raster::raster(matrix(runif(100), 10, 10))))
img_f <- df2img(df_img, template_raster)
}
