% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{df_samples_creation}
\alias{df_samples_creation}
\title{Create a Data Frame of Samples from Attribute Vectors and Classes}
\usage{
df_samples_creation(vector_attributes, classes, csv_name = "df_samples.csv")
}
\arguments{
\item{vector_attributes}{A list of vectors, where each vector contains attributes for a particular class.}

\item{classes}{A vector of class labels corresponding to each attribute vector in `vector_attributes`.}

\item{csv_name}{A string specifying the name of the CSV file to save the data frame. Default is "df_samples.csv". (Currently not used, as the write operation is commented out)}
}
\value{
A data frame with combined attributes and class labels.
}
\description{
This function creates a data frame by combining attribute vectors with corresponding classes.
}
\examples{
attributes1 <- c(1.1, 2.2, 3.3)
attributes2 <- c(4.4, 5.5, 6.6)
vector_attributes <- list(attributes1, attributes2)
classes <- c("Class1", "Class2")
df_samples <- df_samples_creation(vector_attributes, classes)
}
