% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{downsample_data}
\alias{downsample_data}
\title{Downsample Data Samples Based on a Proportion Value}
\usage{
downsample_data(df_samples, ratio = 1)
}
\arguments{
\item{df_samples}{A data frame containing samples with a column named 'Class' indicating the class of each sample.}

\item{ratio}{A numeric value indicating the proportion of the smallest class size to which other classes should be downsampled. Default is 1.}
}
\value{
A data frame with downsampled samples for each class.
}
\description{
This function performs downsampling on a data frame of samples to ensure that the number of samples per class does not exceed a specified proportion of the smallest class.
}
\examples{
df_samples <- data.frame(Class = rep(c("Class1", "Class2"), each = 100), Value = rnorm(200))
downsampled_df <- downsample_data(df_samples, ratio = 0.5)
}
