% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{exhaustive_search_jm_optimal}
\alias{exhaustive_search_jm_optimal}
\title{Exhaustive Search for Optimal Feature Set Based on Jeffries-Matusita Distance}
\usage{
exhaustive_search_jm_optimal(df_samples, classes, all_features)
}
\arguments{
\item{df_samples}{A data frame containing samples with a column named 'Class' indicating the class of each sample.}

\item{classes}{A vector of class labels to include in the distance calculation.}

\item{all_features}{A vector of all possible feature names to consider in the search.}
}
\value{
A list containing two data frames:
\itemize{
  \item \code{distances}: A data frame with each feature set and their corresponding mean Jeffries-Matusita distance.
  \item \code{results}: A data frame with the optimal feature sets, their gains, and mean Jeffries-Matusita distances.
}
}
\description{
This function performs an exhaustive search to find the optimal set of features that maximizes the mean Jeffries-Matusita distance between classes in a data frame of samples.
}
\examples{
df_samples <- data.frame(Class = rep(c("Class1", "Class2", "Class3"), each = 100),
                         Feature1 = rnorm(300), Feature2 = rnorm(300), Feature3 = rnorm(300))
classes <- c("Class1", "Class2", "Class3")
all_features <- c("Feature1", "Feature2", "Feature3")
result <- exhaustive_search_jm_optimal(df_samples, classes, all_features)
print(result)
}
