% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{gamma_bhattacharyya_distance}
\alias{gamma_bhattacharyya_distance}
\title{Calculate Gamma Bhattacharyya Distance Between Two Classes}
\usage{
gamma_bhattacharyya_distance(df_samples, class1, class2, L)
}
\arguments{
\item{df_samples}{A data frame containing samples with a column named 'Class' indicating the class of each sample.}

\item{class1}{A character string specifying the label of the first class.}

\item{class2}{A character string specifying the label of the second class.}

\item{L}{A numeric value used in the calculation of the Bhattacharyya distance.}
}
\value{
A numeric value representing the Bhattacharyya distance between the two classes.
}
\description{
This function calculates the Bhattacharyya distance between two classes in a data frame of samples using a Gamma distribution.
}
\examples{
df_samples <- data.frame(Class = rep(c("Class1", "Class2"), each = 100),
                         Feature1 = rgamma(200, shape = 2, rate = 1))
distance <- gamma_bhattacharyya_distance(df_samples, "Class1", "Class2", L = 2)
}
