% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{gamma_class_likelihood}
\alias{gamma_class_likelihood}
\title{Calculate Gamma Class Likelihood for a Data Frame of Samples}
\usage{
gamma_class_likelihood(dataframe, classes_parameters)
}
\arguments{
\item{dataframe}{A data frame where each row represents a sample and each column represents a feature.}

\item{classes_parameters}{A list of class parameters, where each element is a list containing alpha and beta for a class.}
}
\value{
A data frame of Gamma likelihoods for each sample, with one column per class.
}
\description{
This function calculates the Gamma likelihood for each sample in a data frame based on the provided class parameters (alpha and beta).
}
\examples{
dataframe <- data.frame(Feature1 = rgamma(100, shape = 2, rate = 1), Feature2 = rgamma(100, shape = 2, rate = 1))
class1_params <- list(alpha = 2, beta = 1)
class2_params <- list(alpha = 2, beta = 1.5)
classes_parameters <- list(class1_params, class2_params)
likelihood <- gamma_class_likelihood(dataframe, classes_parameters)
print(likelihood)
}
