% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_fitting_test_KS.R
\name{gamma_euler_ks_test}
\alias{gamma_euler_ks_test}
\title{Gamma Euler Distribution Kolmogorov-Smirnov Test}
\usage{
gamma_euler_ks_test(
  data,
  significance_level = 0.05,
  method = "mle",
  plot = FALSE
)
}
\arguments{
\item{data}{A numeric vector of data to be tested.}

\item{significance_level}{A numeric value representing the significance level for the test. Default is 0.05.}

\item{method}{A character string specifying the method used for parameter estimation. Default is 'mle' (maximum likelihood estimation).}

\item{plot}{A logical value indicating whether to plot the histogram with the theoretical Gamma distribution curve. Default is FALSE.}
}
\value{
A list containing the estimated parameters (alpha and beta) and the p-value of the Kolmogorov-Smirnov test.
}
\description{
This function performs a Kolmogorov-Smirnov test to check if the data follows Euler's Gamma distribution, estimates the parameters of the Gamma distribution, and optionally plots the histogram with the theoretical curve.
}
\examples{
library(fitdistrplus)
library(ggplot2)
data <- rgamma(100, shape = 2, rate = 1)
result <- gamma_euler_ks_test(data, significance_level = 0.05, method = 'mle', plot = TRUE)
print(result)
}
