% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_fitting_test_KS.R
\name{gamma_fitting}
\alias{gamma_fitting}
\title{Fit Gamma Distribution to Sample Classes and Perform KS Test}
\usage{
gamma_fitting(
  df_samples,
  classes,
  significance_level = 0.05,
  method = "mle",
  plot = FALSE
)
}
\arguments{
\item{df_samples}{A data frame containing samples with a column named 'Class' indicating the class of each sample.}

\item{classes}{A vector of class labels to include in the fitting process.}

\item{significance_level}{A numeric value representing the significance level for the KS test. Default is 0.05.}

\item{method}{A character string specifying the method used for parameter estimation in the Gamma distribution. Default is 'mle' (maximum likelihood estimation).}

\item{plot}{A logical value indicating whether to plot the histogram with the theoretical Gamma distribution curve for each class. Default is FALSE.}
}
\value{
A data frame with the results of the Gamma fitting and KS test for each class, including the estimated parameters (Alpha and Beta), p-value, and a decision to accept or reject the null hypothesis.
}
\description{
This function fits a Gamma distribution to each class in a data frame of samples, performs a Kolmogorov-Smirnov test, and optionally plots the results.
}
\examples{
library(fitdistrplus)
library(ggplot2)
df_samples <- data.frame(Class = rep(c("Class1", "Class2"), each = 100), Value = rgamma(200, shape = 2, rate = 1))
classes <- c("Class1", "Class2")
result_df <- gamma_fitting(df_samples, classes, significance_level = 0.05, method = 'mle', plot = FALSE)
print(result_df)
}
