% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_fitting_test_KS.R
\name{gamma_fitting_specific}
\alias{gamma_fitting_specific}
\title{Specific Gamma Fitting for Given ENL}
\usage{
gamma_fitting_specific(
  df_samples,
  classes,
  best_enl,
  significance_level = 0.05
)
}
\arguments{
\item{df_samples}{A data frame containing samples with a column named 'Class' indicating the class of each sample.}

\item{classes}{A vector of class labels to include in the fitting process.}

\item{best_enl}{A numeric value representing the best ENL to be used for fitting.}

\item{significance_level}{A numeric value representing the significance level for the KS test. Default is 0.05.}
}
\value{
A data frame with the results of the specific Gamma fitting for each class, including the shape and rate parameters, p-value, and a decision to accept or reject the null hypothesis.
}
\description{
This function performs a specific fitting of the Gamma distribution for a given Equivalent Number of Looks (ENL) to each class in a data frame of samples.
}
\examples{
library(fitdistrplus)
library(ggplot2)
df_samples <- data.frame(Class = rep(c("Class1", "Class2"), each = 100), Value = rgamma(200, shape = 2, rate = 1))
classes <- c("Class1", "Class2")
best_enl <- 2
result_df_specific <- gamma_fitting_specific(df_samples, classes, best_enl, significance_level = 0.05)
print(result_df_specific)
}
