% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{gamma_parameters_estimation}
\alias{gamma_parameters_estimation}
\title{Estimate Gamma Parameters for a Matrix of Samples}
\usage{
gamma_parameters_estimation(matrix, ENL)
}
\arguments{
\item{matrix}{A numeric matrix where each row represents a sample and each column represents a feature.}

\item{ENL}{A numeric value representing the Equivalent Number of Looks (ENL).}
}
\value{
A list containing the estimated alpha and beta parameters.
}
\description{
This function estimates the parameters of a Gamma distribution (alpha and beta) for a given matrix of samples.
}
\examples{
matrix <- matrix(rgamma(200, shape = 2, rate = 1), nrow = 100, ncol = 2)
ENL <- 2
params <- gamma_parameters_estimation(matrix, ENL)
print(params)
}
