% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{gaussian_bhattacharyya_distance}
\alias{gaussian_bhattacharyya_distance}
\title{Calculate Gaussian Bhattacharyya Distance Between Two Classes}
\usage{
gaussian_bhattacharyya_distance(df_samples, class1, class2)
}
\arguments{
\item{df_samples}{A data frame containing samples with a column named 'Class' indicating the class of each sample.}

\item{class1}{A character string specifying the label of the first class.}

\item{class2}{A character string specifying the label of the second class.}
}
\value{
A numeric value representing the Bhattacharyya distance between the two classes.
}
\description{
This function calculates the Bhattacharyya distance between two classes in a data frame of samples using a Gaussian distribution.
}
\examples{
df_samples <- data.frame(Class = rep(c("Class1", "Class2"), each = 100),
                         Feature1 = rnorm(200), Feature2 = rnorm(200))
distance <- gaussian_bhattacharyya_distance(df_samples, "Class1", "Class2")
}
