% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{gaussian_class_likelihood}
\alias{gaussian_class_likelihood}
\title{Calculate Gaussian Class Likelihood for a Data Frame of Samples}
\usage{
gaussian_class_likelihood(dataframe, classes_parameters)
}
\arguments{
\item{dataframe}{A data frame where each row represents a sample and each column represents a feature.}

\item{classes_parameters}{A list of class parameters, where each element is a list containing a mean vector and covariance matrix for a class.}
}
\value{
A data frame of Gaussian likelihoods for each sample, with one column per class.
}
\description{
This function calculates the Gaussian likelihood for each sample in a data frame based on the provided class parameters (mean vector and covariance matrix).
}
\examples{
dataframe <- data.frame(Feature1 = rnorm(100), Feature2 = rnorm(100))
class1_params <- list(mean = colMeans(dataframe), covariance_matrix = cov(dataframe))
class2_params <- list(mean = colMeans(dataframe + 1), covariance_matrix = cov(dataframe + 1))
classes_parameters <- list(class1_params, class2_params)
likelihood <- gaussian_class_likelihood(dataframe, classes_parameters)
print(likelihood)
}
