% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{gaussian_parameters_estimation}
\alias{gaussian_parameters_estimation}
\title{Estimate Gaussian Parameters for a Matrix of Samples}
\usage{
gaussian_parameters_estimation(matrix)
}
\arguments{
\item{matrix}{A numeric matrix where each row represents a sample and each column represents a feature.}
}
\value{
A list containing the estimated mean vector and covariance matrix.
}
\description{
This function estimates the parameters of a Gaussian distribution (mean vector and covariance matrix) for a given matrix of samples.
}
\examples{
matrix <- matrix(rnorm(200), nrow = 100, ncol = 2)
params <- gaussian_parameters_estimation(matrix)
print(params)
}
