% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{img2df}
\alias{img2df}
\title{Convert an Image Array to a Data Frame}
\usage{
img2df(image, band_names)
}
\arguments{
\item{image}{A 3D array representing the image with dimensions (rows, columns, bands).}

\item{band_names}{A character vector with the names of the bands.}
}
\value{
A data frame where each column corresponds to a band in the image.
}
\description{
This function converts a 3D image array into a data frame, where each column represents a band.
}
\examples{
image <- array(runif(300), dim = c(10, 10, 3))
band_names <- c("Band1", "Band2", "Band3")
df_img <- img2df(image, band_names)

}
