% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{intensity_joint_likelihood}
\alias{intensity_joint_likelihood}
\title{Calculate Joint Intensity Likelihood for a Pair of Images}
\usage{
intensity_joint_likelihood(dataframe, class_parameters, enl)
}
\arguments{
\item{dataframe}{A data frame where each row represents a sample and each column represents a feature.}

\item{class_parameters}{A list of class parameters, where each element is a list containing mu1, mu2, and ro2 for a class.}

\item{enl}{A numeric value representing the Equivalent Number of Looks (ENL).}
}
\value{
A data frame of joint intensity likelihoods for each sample, with one column per class.
}
\description{
This function calculates the joint intensity likelihood for a pair of images based on the provided class parameters.
}
\examples{
dataframe <- data.frame(Image1 = rgamma(100, shape = 2, rate = 1), Image2 = rgamma(100, shape = 2, rate = 1))
class1_params <- list(mu1 = mean(dataframe$Image1), mu2 = mean(dataframe$Image2), ro2 = 0.5)
class_parameters <- list(class1_params)
enl <- 2
likelihood <- intensity_joint_likelihood(dataframe, class_parameters, enl)
print(likelihood)
}
