% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{intensity_joint_parameters_estimation}
\alias{intensity_joint_parameters_estimation}
\title{Estimate Joint Intensity Parameters for a Pair of Images}
\usage{
intensity_joint_parameters_estimation(df)
}
\arguments{
\item{df}{A data frame containing two columns, each representing an image.}
}
\value{
A list containing the estimated parameters: ro2, mu1, and mu2.
}
\description{
This function estimates the joint intensity parameters (mu1, mu2, ro2) for a pair of images.
}
\examples{
df <- data.frame(Image1 = rnorm(100), Image2 = rnorm(100))
params <- intensity_joint_parameters_estimation(df)
print(params)
}
