% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_fitting_test_KS.R
\name{iterative_gamma_fitting}
\alias{iterative_gamma_fitting}
\title{Iterative Gamma Fitting to Find the Best ENL}
\usage{
iterative_gamma_fitting(
  df_samples,
  classes,
  significance_level = 0.05,
  method = "mle"
)
}
\arguments{
\item{df_samples}{A data frame containing samples with a column named 'Class' indicating the class of each sample.}

\item{classes}{A vector of class labels to include in the fitting process.}

\item{significance_level}{A numeric value representing the significance level for the KS test. Default is 0.05.}

\item{method}{A character string specifying the method used for parameter estimation in the Gamma distribution. Default is 'mle' (maximum likelihood estimation).}
}
\value{
A list containing the best ENL, the fitting result for the best ENL, and a data frame with ENL and acceptance count.
}
\description{
This function performs iterative fitting of the Gamma distribution to find the best Equivalent Number of Looks (ENL) based on the results of a Kolmogorov-Smirnov test. It returns the best ENL along with the fitting results.
}
\examples{
library(fitdistrplus)
library(ggplot2)
df_samples <- data.frame(Class = rep(c("Class1", "Class2"), each = 100), Value = rgamma(200, shape = 2, rate = 1))
classes <- c("Class1", "Class2")
result <- iterative_gamma_fitting(df_samples, classes, significance_level = 0.05, method = 'mle')
print(result)
}
