% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{jeffryes_matusita_distance_matrix}
\alias{jeffryes_matusita_distance_matrix}
\title{Calculate Jeffries-Matusita Distance Matrix for All Pairs of Classes}
\usage{
jeffryes_matusita_distance_matrix(df_samples, classes)
}
\arguments{
\item{df_samples}{A data frame containing samples with a column named 'Class' indicating the class of each sample.}

\item{classes}{A vector of class labels to include in the distance matrix.}
}
\value{
A distance matrix representing the Jeffries-Matusita distances between all pairs of classes.
}
\description{
This function calculates the Jeffries-Matusita distance matrix for all pairs of classes in a data frame of samples.
}
\examples{
df_samples <- data.frame(Class = rep(c("Class1", "Class2", "Class3"), each = 100),
                         Feature1 = rnorm(300), Feature2 = rnorm(300))
classes <- c("Class1", "Class2", "Class3")
jm_distance_matrix <- jeffryes_matusita_distance_matrix(df_samples, classes)
}
