% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_CMAP_functions.R
\name{masking_replace}
\alias{masking_replace}
\title{Replace Values in an Image Based on a Mask}
\usage{
masking_replace(image, mask, value)
}
\arguments{
\item{image}{A RasterStack or RasterBrick object representing the image.}

\item{mask}{A RasterLayer object representing the mask.}

\item{value}{The value to replace in the image where the mask is TRUE.}
}
\value{
A RasterStack or RasterBrick object with the modified values.
}
\description{
This function replaces values in an image based on a mask, applying the specified value where the mask is TRUE.
}
\examples{
# Carregando o pacote necessário
library(raster)

# Create an example image and a mask
image <- stack(raster(matrix(1:100, 10, 10)), raster(matrix(1:100, 10, 10)))
mask <- raster(matrix(sample(0:1, 100, replace = TRUE), 10, 10))

# Replace values based on the mask
result <- masking_replace(image, mask, value = 0)

# Plot the result
plot(result)
}
