% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{maximum_likelihood_classifier}
\alias{maximum_likelihood_classifier}
\title{Maximum Likelihood Classifier}
\usage{
maximum_likelihood_classifier(likelihoods)
}
\arguments{
\item{likelihoods}{A data frame where each column represents the likelihood of a sample belonging to a specific class.}
}
\value{
A data frame with one column named \code{predicted_class_id}, containing the predicted class for each sample.
}
\description{
This function classifies samples based on the maximum likelihood from a set of likelihoods.
}
\examples{
likelihoods <- data.frame(Class1 = c(0.1, 0.6, 0.3), Class2 = c(0.4, 0.2, 0.5), Class3 = c(0.5, 0.2, 0.2))
predictions <- maximum_likelihood_classifier(likelihoods)
print(predictions)
}
