% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{mean_jm_distance}
\alias{mean_jm_distance}
\title{Calculate Mean Jeffries-Matusita Distance Between All Classes}
\usage{
mean_jm_distance(df_samples, classes)
}
\arguments{
\item{df_samples}{A data frame containing samples with a column named 'Class' indicating the class of each sample.}

\item{classes}{A vector of class labels to include in the distance calculation.}
}
\value{
A numeric value representing the mean Jeffries-Matusita distance between all pairs of classes.
}
\description{
This function calculates the mean Jeffries-Matusita distance between all pairs of classes in a data frame of samples.
}
\examples{
df_samples <- data.frame(Class = rep(c("Class1", "Class2", "Class3"), each = 100),
                         Feature1 = rnorm(300), Feature2 = rnorm(300))
classes <- c("Class1", "Class2", "Class3")
mean_distance <- mean_jm_distance(df_samples, classes)
print(mean_distance)
}
