% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_fitting_test_KS.R
\name{normal_fitting}
\alias{normal_fitting}
\title{Fit Normal Distribution to Sample Classes and Perform KS Test}
\usage{
normal_fitting(
  df_samples,
  classes,
  significance_level = 0.05,
  method = "mle",
  plot = FALSE
)
}
\arguments{
\item{df_samples}{A data frame with sample data.}

\item{classes}{A vector of class labels.}

\item{significance_level}{The significance level for the KS test.}

\item{method}{The method for fitting the distribution (default is 'mle').}

\item{plot}{Logical. If TRUE, plots the fitted distribution.}
}
\value{
A data frame with the fitting results.
}
\description{
This function fits a normal distribution to sample classes and performs the KS test.
}
\examples{
library(fitdistrplus)
library(ggplot2)
df_samples <- data.frame(Class = rep(c("Class1", "Class2"), each = 100), Value = rnorm(200))
classes <- c("Class1", "Class2")
result_df <- normal_fitting(df_samples, classes, significance_level = 0.05, method = 'mle', plot = TRUE)
}
