% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_fitting_test_KS.R
\name{normal_fitting_mv}
\alias{normal_fitting_mv}
\title{Fit Multivariate Normal Distribution to Sample Classes and Perform Mardia Test}
\usage{
normal_fitting_mv(df_samples, classes, significance_level = 0.05)
}
\arguments{
\item{df_samples}{A data frame containing samples with a column named 'Class' indicating the class of each sample.}

\item{classes}{A vector of class labels to include in the fitting process.}

\item{significance_level}{A numeric value representing the significance level for the Mardia test. Default is 0.05.}
}
\value{
A data frame with the results of the Multivariate Normal fitting and Mardia test for each class, including the p-values for skewness and kurtosis, and a decision to accept or reject the null hypothesis.
}
\description{
This function fits a Multivariate Normal distribution to each class in a data frame of samples and performs Mardia's test for multivariate normality.
}
\examples{
library(MVN)
df_samples <- data.frame(Class = rep(c("Class1", "Class2"), each = 100),
                         Feature1 = rnorm(200), Feature2 = rnorm(200))
classes <- c("Class1", "Class2")
result_df <- normal_fitting_mv(df_samples, classes, significance_level = 0.05)
print(result_df)
}
