% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_fitting_test_KS.R
\name{normal_ks_test}
\alias{normal_ks_test}
\title{Perform Kolmogorov-Smirnov Test for Normality}
\usage{
normal_ks_test(data, significance_level = 0.05, method = "mle")
}
\arguments{
\item{data}{A numeric vector of data to be tested.}

\item{significance_level}{A numeric value representing the significance level for the test. Default is 0.05.}

\item{method}{A character string specifying the method used for parameter estimation. Default is 'mle' (maximum likelihood estimation).}
}
\value{
A list containing the estimated mean, standard deviation, and the p-value of the KS test.
}
\description{
This function performs a Kolmogorov-Smirnov test to check if the data follows a Normal distribution and estimates the parameters.
}
\examples{
data <- rnorm(100)
result <- normal_ks_test(data, significance_level = 0.05, method = 'mle')
print(result)
}
