% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_fitting_test_KS.R
\name{normal_mardia_test}
\alias{normal_mardia_test}
\title{Perform Mardia Test for Multivariate Normality}
\usage{
normal_mardia_test(data, significance_level = 0.05)
}
\arguments{
\item{data}{A data frame or matrix of data to be tested.}

\item{significance_level}{A numeric value representing the significance level for the test. Default is 0.05.}
}
\value{
A list containing the p-values for skewness and kurtosis.
}
\description{
This function performs Mardia's test to check if the data follows a Multivariate Normal distribution.
}
\examples{
data <- matrix(rnorm(300), ncol = 3)
result <- normal_mardia_test(data, significance_level = 0.05)
print(result)
}
