% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{parameters_estimation}
\alias{parameters_estimation}
\title{Estimate Parameters for Each Class in a Data Frame of Samples}
\usage{
parameters_estimation(df_samples, classes, p_function = "gaussian", ENL = NULL)
}
\arguments{
\item{df_samples}{A data frame containing samples with a column named 'Class' indicating the class of each sample.}

\item{classes}{A vector of class labels to include in the parameter estimation.}

\item{p_function}{A character string specifying the probability distribution function to be used ('gaussian', 'gamma', or 'intensity_joint_distribution'). Default is 'gaussian'.}

\item{ENL}{An optional numeric value used in the estimation for the 'gamma' distribution.}
}
\value{
A list of parameter estimates for each class.
}
\description{
This function estimates parameters for each class in a data frame of samples using different probability distribution functions.
}
\examples{
df_samples <- data.frame(Class = rep(c("Class1", "Class2"), each = 100),
                         Feature1 = rnorm(200), Feature2 = rnorm(200))
classes <- c("Class1", "Class2")
params_list <- parameters_estimation(df_samples, classes, p_function = 'gaussian')
}
