% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{plot_dendrogram}
\alias{plot_dendrogram}
\title{Plot Dendrogram Based on Distance Matrix}
\usage{
plot_dendrogram(dist_matrix, threshold = 1, method = "ward.D2", name = NULL)
}
\arguments{
\item{dist_matrix}{A distance matrix used for hierarchical clustering.}

\item{threshold}{A numeric value specifying the threshold for cutting the dendrogram. Default is 1.}

\item{method}{A character string specifying the clustering method. Default is "ward.D2".}

\item{name}{A character string specifying the name/title of the dendrogram plot.}
}
\value{
An object of class \code{hclust} which describes the tree produced by the clustering process.
}
\description{
This function generates a dendrogram from a previously calculated distance matrix.
}
