% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_confusion_matrix.R
\name{replace_class_names}
\alias{replace_class_names}
\title{Replace Class Names}
\usage{
replace_class_names(data, column, dictionary)
}
\arguments{
\item{data}{A dataframe containing the column whose values will be replaced.}

\item{column}{A string specifying the name of the column whose values will be replaced.}

\item{dictionary}{A named vector where the names represent the old values and the values represent the new class names.}
}
\value{
The dataframe with the specified column containing the new class names.
}
\description{
This function replaces the class names in a specific column of a dataframe according to a provided dictionary.
}
\examples{
data <- data.frame(Class = c("A", "B", "C"))
dictionary <- c("A" = "Class1", "B" = "Class2", "C" = "Class3")
data <- replace_class_names(data, "Class", dictionary)
}
