% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{split_classes}
\alias{split_classes}
\title{Split Samples into Class-specific Matrices}
\usage{
split_classes(df_samples, classes)
}
\arguments{
\item{df_samples}{A data frame containing samples with a column named 'Class' indicating the class of each sample.}

\item{classes}{A vector of class labels to split the samples into.}
}
\value{
A list of matrices, each containing the samples for a specific class.
}
\description{
This function splits a data frame of samples into separate matrices for each specified class.
}
\examples{
df_samples <- data.frame(Class = rep(c("Class1", "Class2"), each = 5),
                        Feature1 = rnorm(10), Feature2 = rnorm(10))
classes <- c("Class1", "Class2")
class_matrices <- split_classes(df_samples, classes)
}
