% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_auxiliary_functions.R
\name{split_samples}
\alias{split_samples}
\title{Split Samples into Training and Validation Sets}
\usage{
split_samples(df_samples, ratio, samples_per_class = NULL)
}
\arguments{
\item{df_samples}{A data frame containing samples with a column named 'Class' indicating the class of each sample.}

\item{ratio}{A numeric value indicating the proportion of the data to include in the training set. Ignored if `samples_per_class` is provided.}

\item{samples_per_class}{An optional integer specifying the fixed number of samples to use for training from each class. If provided, `ratio` is ignored.}
}
\value{
A list with two elements: `train_samples`, a data frame of training samples, and `valid_samples`, a data frame of validation samples.
}
\description{
This function splits a data frame of samples into training and validation sets based on a specified ratio or a fixed number of samples per class.
}
\examples{
df_samples <- data.frame(Class = rep(c("Class1", "Class2"), each = 100),
                         Feature1 = rnorm(200), Feature2 = rnorm(200))
# Split with a ratio
split_result <- split_samples(df_samples, ratio = 0.7)
train_samples <- split_result$train_samples
valid_samples <- split_result$valid_samples

# Split with a fixed number of samples per class
split_result_fixed <- split_samples(df_samples, samples_per_class = 50)
train_samples_fixed <- split_result_fixed$train_samples
valid_samples_fixed <- split_result_fixed$valid_samples
}
