% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_CMAP_functions.R
\name{transition_matrix}
\alias{transition_matrix}
\title{Create Transition Matrix}
\usage{
transition_matrix(file_paths, output)
}
\arguments{
\item{file_paths}{A character vector containing the file paths of the input transition matrices.}

\item{output}{A character string specifying the file path for the output transition matrix.}
}
\value{
A data frame representing the combined transition matrix.
}
\description{
This function creates a transition matrix from multiple input files and writes the result to an output file.
}
\examples{
# Assuming example transition matrices are saved in "trans1.txt", "trans2.txt"
write.table(matrix(c(0.1, 0.9, 0.3, 0.7), nrow = 2), "trans1.txt", sep = ";", col.names = FALSE, row.names = FALSE)
write.table(matrix(c(0.8, 0.2, 0.6, 0.4), nrow = 2), "trans2.txt", sep = ";", col.names = FALSE, row.names = FALSE)
file_paths <- c("trans1.txt", "trans2.txt")
output <- "combined_transitions.txt"
transition_matrix(file_paths, output)
}
