% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_CMAP_functions.R
\name{verify_dimensions}
\alias{verify_dimensions}
\title{Verify if All Images Have the Same Dimensions}
\usage{
verify_dimensions(paths)
}
\arguments{
\item{paths}{A character vector containing the file paths of the images to be checked.}
}
\value{
A logical value indicating whether all images have the same dimensions (TRUE) or not (FALSE).
}
\description{
This function checks if all images in the provided file paths have the same dimensions.
}
