% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_CMAP_functions.R
\name{compareRasterClassification}
\alias{compareRasterClassification}
\title{Compare Raster Classifications}
\usage{
compareRasterClassification(baseDirectory_1, baseDirectory_2, n, maskPath)
}
\arguments{
\item{baseDirectory_1}{The base directory 1}

\item{baseDirectory_2}{The base directory 2}

\item{n}{Number of images.}

\item{maskPath}{The path to the cloud mask raster file.}
}
\value{
A list containing the accumulated differences and the final percentage of disagreement.
}
\description{
This function compares raster classifications from two sources over multiple years and calculates the differences.
}
\examples{
# Assuming example classification rasters are saved in "ML/Classification_2000.TIF", "CMAP/CMAP_1.TIF"
baseDirectory_ML <- "path/to/ML/"
baseDirectory_CMAP <- "path/to/CMAP/"
years <- c(2000, 2001, 2002)
maskPath <- "path/to/cloud_mask.TIF"
if (file.exists(maskPath) && all(file.exists(paste0(baseDirectory_1, "Classification_", n, ".TIF"))) && all(file.exists(paste0(baseDirectory_2, "classification_", 1:length(n), ".TIF")))) {
  result <- compareRasterClassification(baseDirectory_1, baseDirectory_2, years, maskPath)
} else {
  message("Example files do not exist. Please provide the correct paths to the classification rasters and cloud mask.")
}
}
